fullstack-app/backend/src/backend/database/connection.py
import os
from dotenv import load_dotenv
from sqlmodel import create_engine, Session

# Load semua isi file .env ke dalam sistem
load_dotenv()

# Ambil variabel DATABASE_URL dari file .env
DATABASE_URL = os.getenv("DATABASE_URL")

engine = create_engine(DATABASE_URL, echo=True)

def get_session():
    with Session(engine) as session:
        yield session