fullstack-app/backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
authors = [
    {name = "trishagarniss",email = "trishagarniss@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "sqlmodel (>=0.0.37,<0.0.38)",
    "sqlalchemy (>=2.0.48,<3.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "alembic (>=1.18.4,<2.0.0)",
    "python-dotenv (>=1.2.2,<2.0.0)"
]

[tool.poetry]
packages = [{include = "backend", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poe.tasks]
dev = "fastapi dev src/backend/app.py"